UserClass methods added: Show program/reservations, Add/Delete/Edit Movies, Add/Delete/Edit Bookings
Methods to be added:
Main method
Save/Load/Reset Data

Deviations from class diagram:
Names of methods in User() changed to plural for corresponding methods in the admin classes.
displayInfo() returns an int for decision in the main method. This can be changed when the main method is added, so it takes an int that method itself.

Alerts:
toString is to be added in MovieAdmin for the variables, or must come up with a different way in displayMovieProgram().