11.05.2022
Movies with default name are not added in the arraylist. Methods instead check for null objects at all times to prevent exceptions.
editMovies() now has the feature to replace movies in the movie program with movies in the rest of the arraylist. Consequently, loadingData() simply suggests doing this in editMovies() rather than being able to replace movies as soon as data is loaded.
 Success and Error messages are placed in appropriate positions
 Iterating and printing booking is changed, so it prints list.indexOf() instead of an int.